rename groupings module to system module
As I moved the EnergySystem into this module the old name was confusing.
